Refactor Gauge viz to TypeScript and a function component
