+++ frontend/src/metabase/visualizations/visualizations/Gauge/utils.ts
@@ -11,3 +11,7 @@ export const getValue = (rows: unknown[][]) => {
 
   return rawValue;
 };
+
+export const radians = (degrees: number) => (degrees * Math.PI) / 180;
+
+export const degrees = (radians: number) => (radians * 180) / Math.PI;
